[CHG] simplify the onchange of the product template:
When the template is changed, discard all the variants, and add a line per variant of the template.
Do not try to keep the configured quantity or whatever, that's too many code for very little value.
By the way, an onchange must never write on the DB, but just read