The leftovers dispatches are now grouped by the quantity of products in the packs.
This improves the deliveries as even if the products are differents, the size of the packs will likely be the same
in each generated leftover dispatch.

Once all the leftovers dispatches grouped by quantity have been generated, smaller dispatches can be generated for the rest,
we reapply the leftover algorithm on these ones so we finally create 'leftovers of leftovers' with various quantities.