Use an optimized version of setlast_tracking
The original version grows exponentially because it loop over
all the moves of the picking to gather the trackings, only to keep
the last one.

To split one move of 1000 lines in 1000 moves of 1 lines, it takes 90
seconds instead of 330 on my computer.

No gain on a move with 100 lines or 10 lines though.