Explicit --without-demo in marabunta install_args
This is a workaround for https://github.com/odoo/odoo/pull/12730
The option in the odoo configuration file is not considered, but the CLI
option always is. So we force the command line to always disable demo
data. The drawback is that the DEMO flag won't work. This line should be
removed when the fix is merged.