Rework connector's shipping bindings
* Remove QoQa Shipper Services
* Rename QoQa Shipper Rates to QoQa Shipper Fees
* QoQa Package types are now "delivery.carrier"

Fees are only used on sales order lines: we use the name and the
product_id for the shipping order line.

Package types are used for the export of packages to the API, it
indicates the method used for the delivery. We'll probably have more
methods on Odoo than on the QoQa BO: some methods on Odoo (Vinolog X,
Vinolog Y, Vinolog Z) might be all linked with "Vinolog" on the BO.