Fix locations complete names
Triggering the complete_name computation on every records did not work
for all the records, because it was dependent of the order of execution.
A complete name could be constructed with the parent name which could
not be correct yet. When we trigger the change only on the parent nodes
of the tree, the trigger will be propagated down the leaves.