Add addon that logs some requests info
It logs a selection of the info of the requests and the responses, those
which can be useful to analyze the usage and duration of requests.

The logs are output as JSON, so they are easier to parse afterwards when
they'll be loaded in logstash / elasticseach / kibana.

List of things for the future:

* review the fields we need
* review the fields names
* some fields / data containers are not always present, ensure that we
  don't fail when one is missing
* see if we need to use syslog rather than stdout for the ELK stack
* filter the requests we want to analyze