Follow binding relations even on inactive records
When a record, such as a 'sale.order' is set as 'active=False', its
binding will inherit this value with the effect that when calling
'sale.qoqa_bind_ids', nothing will be returned.

Usually, we expect the binding to be returned even when the record is
inactive.  This was causing many small issues in the code when we loop
over qoqa_bind_ids.

Now, the relation will be read with a 'active_test=False' context, so
will see the binding whatever the active value is on the record.