Hide parent name for QoQa customers
The parent name is the login and email of the client when they are
imported from the QoQa webshop. We never ever want them to be displayed
on reports. So the qweb widgets sets a flag in the context and a special
case is introduced in name_get to return only the real name of the
client (which is the name on the address, not the main partner).

The context flag also allow to remove the override of
Contact.record.to_html, as it was there to handle the ',' added in the
name. Now, we just don't return any comma in the name.