[WIP] Prepare migration of QoQa addresses
We don't have the file from QoQa yet, hence the 'WIP' flag.

During their migration on QoQa4, they will duplicate all the addresses
linked to a sales order, we have to follow this change. This migration
script duplicates the addresses and create the new binding according to
a mapping file.