Add various index
The rationale for adding them is that they are all for queries which
appeared in the "Normalized slowest queries" section of pgbadger for
Postgres logs registered during a browsing session. This browsing
session only involved some basic browsing of the most common views
(display of partners, sales orders, pickings, dispatches, ... list and
form views).