Add a script to move back S3 small files to DB
The access to a file stored on the objects storage is slower
than a local disk or database access. For attachments like
image_small that are accessed in batch for kanban views, this
is too slow. We store this type of attachment in the database.

This method will only be used once during the migration to Odoo 9.0

It is copied from a migration of the 'attachment_s3' module.

This migration should be run once the production is running by running
scripts/migrate_attachment_image_small.py