Add random sleep when jobs export the same record
When we export a template with more than 1 variant, we see jobs being
stalled because we can export only 1 variant at a time, and the jobs are
pushed back and retried at the same time. With a more random retry
within 5 to 20 seconds, they have less chances to be exported at the
same time and to hit a RetryableJobError again.