Improve performance for validation of refunds
The issue is explained on https://github.com/odoo/odoo/pull/18970
the _get_sale_order_line method loads a lot of records in the prefetch,
which are not needed later. As soon as we'll try to read an attribute of
one of the model added by this method, all the records in prefetch will
be read and will slow a lot the validation of the refund.

BSQOQ-16 BIZ-109