Import vouchers as new order lines
Previously, they were imported as move lines of payment.
Now, the vouchers will be negative lines in the sales orders.
This is a special case for the import as on the QoQa side, they are
not represented as a line, so we create a new line from it on the Odoo side
without using the 'qoqa.sale.order.line' mapper.

BIZ-712