Adapt the sale exception rules with vouchers
Vouchers are negative lines on Odoo, thus they modify the total of the order.
The total of the order on the QoQa4 side does not include the vouchers.
Then, usually you'll have something such as:

|          | QoQa4 | Odoo |
| Products |   100 |  100 |
| Voucher  |     0 |  -50 |
| Total    |   100 |   50 |

The new computed field amount_total_without_voucher includes the voucher so in
this example the amount_total_without_voucher will be 100.- and the rule
will be valid.