Support voucher used partially
Voucher (gift cards) can be used only partially. Reading the amount
of the voucher is therefore not correct. The amount to deduce in the sale order
is the one we found in the payment.
We have no way to know which voucher is related to which payment though, so if
we have more than one voucher in the same order, the amount won't be correct.