Return missing_error if the order doesn't exist
When the API for changing the address of a sales order is called and the
sales order doesn't exist, it returns a missing_error. Previously it was
a user_error which was indistinguishable from the error when you can't
change it because it's delivered. The API's client needs it to know that the
sales orders has not yet been imported into odoo.