account_mass_reconcile advanced.ref
Added the 'mass.reconcile.advanced.label' model to at least be able to
reconcile payments as Odoo would. In Odoo the customer invoice has the
payment reference in the account move ref field, and the payment
typically has this in the move line name field. Since the invoice
account move line that we want to reconcile for customer invoices is a
debit line, and the payment we want to find is a credit line, we need to
invert the current implementation of the reconciliation

'mass.reconcile.advanced.ref' should be used to reconcile refunds.