[FIX] account_reconcilication_widget: notifications
Notifications displayed at the start of the reconciliation interface are
back up and running (it was working in v10, broken in v12 and v13).
For example, when you import an OFX file and go directly to the
reconciliation interface, it will display the number of transactions
that were automatically dropped because they were already present in
Odoo (detected via the unique_import_id).